feat(content-sidebar): Rename component to MetadataSidebarRedesign
