+++ src/elements/content-sidebar/MetadataSidebarRedesign.scss
